+++ src/Servidor/Modelo/fireman.h
@@ -11,7 +11,7 @@ class StrategyMoverFireman;
 class Fireman: public Personaje_npc{
 	private:
 		Mapa *mapa;
-		size_t tiempo_pasado;
+		float tiempo_pasado;
 		Arma_fireman* arma;
 		bool atacando;
 		StrategyMoverFireman movimiento;
@@ -22,11 +22,11 @@ class Fireman: public Personaje_npc{
 		//Dado una Mapa* y una direccion, ataca.
 		virtual void atacar(int dir, Mapa* mapa);
 		//Dado un tiempo pasado y un Mapa*, se mueve.
-		virtual void mover(size_t tiempo, Mapa* mapa);
+		virtual void mover(float tiempo, Mapa* mapa);
 		//Dada una Bala*, recibe un ataque.
 		virtual void recibir_ataque(Bala* ataque);
 		//Dado un tiempo, realiza un update.
-		virtual void update(size_t tiempo);
+		virtual void update(float tiempo);
 
 		friend class StrategyMoverFireman;
 };
